chore: split Sidebar into SidebarHeader, SidebarNavigation, SidebarBooks, SidebarPrompts
514-line component extracted into 4 focused files; Sidebar.tsx reduced to ~100-line orchestrator. Pure extraction — no logic changes.
